Use mock class in test
