+++ tests/mr/device/aligned_mr_tests.cpp
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
+#include "../../mock_resource.hpp"
 #include <rmm/detail/aligned.hpp>
 #include <rmm/detail/error.hpp>
 #include <rmm/mr/device/aligned_resource_adaptor.hpp>
@@ -27,16 +28,6 @@ namespace rmm::test {
 namespace {
 
 using ::testing::Return;
-
-class mock_resource : public rmm::mr::device_memory_resource {
- public:
-  MOCK_METHOD(bool, supports_streams, (), (const, override, noexcept));
-  MOCK_METHOD(bool, supports_get_mem_info, (), (const, override, noexcept));
-  MOCK_METHOD(void*, do_allocate, (std::size_t, cuda_stream_view), (override));
-  MOCK_METHOD(void, do_deallocate, (void*, std::size_t, cuda_stream_view), (override));
-  using size_pair = std::pair<std::size_t, std::size_t>;
-  MOCK_METHOD(size_pair, do_get_mem_info, (cuda_stream_view), (const, override));
-};
 
 using aligned_mock = rmm::mr::aligned_resource_adaptor<mock_resource>;
 using aligned_real = rmm::mr::aligned_resource_adaptor<rmm::mr::device_memory_resource>;
